Updated HDF5 deprecated methods to new methods. (Verified by building HDF5 files and then generating the kinship matrix and binary comparing it with a previous version. Note that the HDF5 files show up as different after the update (even though the byte size is still the same and visually using HDFView it appears the same).)
Also added comments to some of the code as well as reformatted the code to make it neater. Added some TODOs as well as uncommented a few lines to make the project buildable (noted by TODOs for easy viewing).

Removed redundant declarations and corrected some spelling mistakes and added some extra text to prompts.

Signed-off-by: Pathum Kossinna <thalagalakossinnagep@ucalgary,ca>